Makefile: adapt to a newer `virtualenv` on Python 2.7
We're not compatible with Python 3 yet, but as a first step let's work
correctly with Python 2.7 with the newest-possible `virtualenv`.

At first I was going to rely on `virtualenv-2.7`, but it turns out
that is something left around by old versions of
`virtualenv`. Instead, invoke `virtualenv` with the '-p' argument to
give which Python interpreter to use for the new virtual environment.